Separate `thread_info` from `thread_control_helper`.
